backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.0
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
prophet==1.1.5
statsmodels==0.14.0
celery==5.3.4
redis==5.0.1
Pillow==10.1.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
python-crontab==2.6.0
